Fix UI and config
The UI and config files weren't being found by the systemd service as it
was looking in the `dataDir` (as that's its working directory). This
ensures they can be seen from there.

Additionally, the `--config` option didn't seem to do anything, but
from browsing the code it looks like `local.toml` is read after the
default file. Without the default file the service didn't start, due to
many missing options.